Fixes for libtatsu
