+++ tsschecker/Makefile.am
@@ -1,5 +1,5 @@
-AM_CFLAGS = $(libplist_CFLAGS) $(libfragmentzip_CFLAGS) $(libcurl_CFLAGS) $(libcrypto_CFLAGS) $(libirecovery_CFLAGS)  -I$(top_srcdir)/external/jssy/jssy/
-AM_LDFLAGS = $(libplist_LIBS) $(libfragmentzip_LIBS) $(libcurl_LIBS) $(libcrypto_LIBS) $(wolfssl_LIBS) $(libirecovery_LIBS) -lm
+AM_CFLAGS = $(libplist_CFLAGS) $(libtatsu_CFLAGS) $(libfragmentzip_CFLAGS) $(libcurl_CFLAGS) $(libcrypto_CFLAGS) $(libirecovery_CFLAGS)  -I$(top_srcdir)/external/jssy/jssy/
+AM_LDFLAGS = $(libplist_LIBS) $(libtatsu_LIBS) $(libfragmentzip_LIBS) $(libcurl_LIBS) $(libcrypto_LIBS) $(wolfssl_LIBS) $(libirecovery_LIBS) -lm
 
 lib_LIBRARIES = libjssy.a
 libjssy_a_CFLAGS = $(AM_CFLAGS)
@@ -7,7 +7,7 @@ libjssy_a_SOURCES = $(top_srcdir)/external/jssy/jssy/jssy.c
 
 if NOMAIN
 noinst_LTLIBRARIES = libtsschecker.la
-libtsschecker_la_SOURCES = tsschecker.c tss.c download.c common.c
+libtsschecker_la_SOURCES = tsschecker.c download.c common.c
 libtsschecker_la_CFLAGS = $(AM_CFLAGS)
 libtsschecker_la_LIBADD = $(AM_LDFLAGS) libjssy.a
 else
@@ -15,7 +15,7 @@ bin_PROGRAMS = tsschecker
 tsschecker_CFLAGS = $(AM_CFLAGS)
 tsschecker_LDADD = $(AM_LDFLAGS) libjssy.a
 
-tsschecker_SOURCES = tsschecker.c tss.c download.c common.c main.c
+tsschecker_SOURCES = tsschecker.c download.c common.c main.c
 endif
 
 if WINDOWS
